декомпозировал визуализацию
